requirements.txt
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
coloredlogs==15.0.1
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.1.6
humanfriendly==10.0
idna==3.11
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mpmath==1.3.0
numpy==2.3.5
onnxruntime==1.23.2
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.0
protobuf==6.33.1
pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
rich==14.2.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
textual==6.6.0
tqdm==4.67.1
typer-slim==0.20.0
typing-extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
PyQt6
